give members new ump tools
