+++ server/static/js/src/entry/account/routes/membership/components/RecurringOrCircle.vue
@@ -1,57 +1,7 @@
 <template>
   <section class="c-detail-box">
     <div class="has-xxl-btm-marg">
-      <p
-        v-if="failureMessage"
-        role="alert"
-        class="has-b-btm-marg has-text-error"
-      >
-        <strong>{{ failureMessage }}</strong>
-      </p>
-      <p
-        v-if="successMessage"
-        role="alert"
-        class="has-b-btm-marg has-text-success"
-      >
-        <strong>{{ successMessage }}</strong>
-      </p>
-      <info-list v-if="!canViewAs" :items="data">
-        <template #text="{ item: { extra, key } }">
-          <template v-if="key === 'donation'">
-            {{ extra.amount | currency }}, {{ extra.period }}
-          </template>
-          <template v-if="key === 'payment'">
-            {{ extra.brand }} ending in {{ extra.last4 }}
-            <div>
-              <button class="has-text-teal" @click="toggleInlinePaymentForm">
-                <span v-if="!openPaymentForm">
-                  Edit
-                  <span class="c-icon c-icon--teal c-icon--baseline t-size-s">
-                    <svg aria-hidden="true"><use href="#pencil-fill"></use></svg>
-                  </span>
-                </span>
-                <span v-if="openPaymentForm">
-                  Close
-                  <span class="c-icon c-icon--teal c-icon--baseline t-size-s">
-                    <svg aria-hidden="true"><use href="#close"></use></svg>
-                  </span>
-                </span>
-              </button>
-            </div>
-            <card-update
-              v-if="openPaymentForm"
-              :stripe-customer-id="extra.stripeCustomerId"
-              @formSubmitted="formSubmitted"
-              @onSuccess="onInlineSuccess"
-              @onFailure="onInlineFailure"
-            ></card-update>
-          </template>
-          <template v-if="key === 'next'">
-            {{ extra.nextTransactionDate | longDate }}
-          </template>
-        </template>
-      </info-list>
-      <table v-if="canViewAs" class="c-table l-width-full">
+      <table class="c-table l-width-full">
         <thead>
           <tr>
             <th class="t-align-left"><strong>Donation</strong></th>
@@ -105,9 +55,8 @@
       </table>
     </div>
     <user-internal-nav show-donation-history />
-    <card-update-new
+    <card-update
       :rdo="stagedRdo"
-      @formSubmitted="formSubmitted"
       @onSuccess="onSuccess"
       @onFailure="onFailure"
       @onClose="onClose('cardModal')" />
@@ -135,30 +84,22 @@ import contextMixin from '../../../store/context/mixin';
 import logError from '../../../utils/log-error';
 import { AxiosError } from '../../../errors';
 
-import ConfirmModal from '../../../components/ConfirmModal.vue';
-import InfoList from '../../../components/InfoList.vue';
-import CardUpdateNew from './CardUpdateNew.vue';
 import CardUpdate from './CardUpdate.vue';
+import ConfirmModal from '../../../components/ConfirmModal.vue';
 import MessageModal from '../../../components/MessageModal.vue';
 
 export default {
   name: 'MembershipRecurringOrCircle',
 
   components: {
-    ConfirmModal,
-    InfoList,
-    CardUpdateNew,
     CardUpdate,
+    ConfirmModal,
     MessageModal,
   },
 
   mixins: [contextMixin, userMixin],
 
   props: {
-    nextTransaction: {
-      type: Object,
-      required: true,
-    },
     recurringTransactions: {
       type: Array,
       required: true,
@@ -171,114 +112,26 @@ export default {
 
   data() {
     return {
-      openPaymentForm: false,
-      openConfirmModal: false,
-      successMessage: '',
-      failureMessage: '',
       confirmHeading: '',
       confirmBody: '',
       messageHeading: '',
       messageBody: '',
       messageType: '',
-      declinedCard: false,
       checkModalResolve: () => {},
       stagedRdo: {},
     }
   },
 
-  computed: {
-    data() {
-      const data = [];
-      const {
-        amount,
-        period,
-        card,
-        stripeCustomerId,
-        date: nextTransactionDate,
-      } = this.nextTransaction;
-
-      data.push({
-        key: 'donation',
-        heading: 'Donation',
-        extra: { amount, period },
-      });
-
-      if (card) {
-        const { brand, last4 } = card;
-
-        data.push({
-          key: 'payment',
-          heading: 'Payment method',
-          extra: { brand, last4, stripeCustomerId },
-        });
-      }
-
-      data.push({
-        key: 'next',
-        heading: 'Next payment',
-        extra: { nextTransactionDate },
-      });
-
-      return data;
-    },
-  },
-
   methods: {
-    toggleInlinePaymentForm() {
-      this.openPaymentForm = !this.openPaymentForm;
-      const gaCardBase = {
-        event: this.ga.customEventName,
-        gaCategory: this.ga.userPortal.category,
-        gaLabel: this.ga.userPortal.labels['update-card'],
-      };
-      if (this.openPaymentForm) {
-        window.dataLayer.push({
-          ...gaCardBase,
-          gaAction: this.ga.userPortal.actions['attempt-card-update'],
-        });
-      } else {
-        window.dataLayer.push({
-          ...gaCardBase,
-          gaAction: this.ga.userPortal.actions['cancel-card-update'],
-        });
-      }
-    },
-
-    onInlineSuccess(message) {
-      this.successMessage = message;
-      this.openPaymentForm = false;
-    },
-  
-    onInlineFailure(message) {
-      this.failureMessage = message;
-      this.openPaymentForm = false;
-    },
-
     togglePaymentForm(rdo) {
-      this.stagedRdo = rdo;
-
-      this.$modal.show('cardModal');
-      const gaCardBase = {
-        event: this.ga.customEventName,
-        gaCategory: this.ga.userPortal.category,
+      window.dataLayer.push({
+        event: 'customUserPortalEditPaymentIntent',
+        gaAction: this.ga.userPortal.actions['attempt-card-update'],
         gaLabel: this.ga.userPortal.labels['update-card'],
-      };
-      if (this.openPaymentForm) {
-        window.dataLayer.push({
-          ...gaCardBase,
-          gaAction: this.ga.userPortal.actions['attempt-card-update'],
-        });
-      } else {
-        window.dataLayer.push({
-          ...gaCardBase,
-          gaAction: this.ga.userPortal.actions['cancel-card-update'],
-        });
-      }
-    },
-  
-    formSubmitted() {
-      this.successMessage = '';
-      this.declinedCard = false;
+      });
+
+      this.stagedRdo = rdo;
+      this.$modal.show('cardModal');
     },
   
     onSuccess(message) {
@@ -302,6 +155,12 @@ export default {
     },
 
     async cancelDonation(rdo) {
+      window.dataLayer.push({
+        event: 'customUserPortalCancelIntent',
+        gaAction: this.ga.userPortal.actions['attempt-cancel-donation'],
+        gaLabel: this.ga.userPortal.labels['cancel-donation'],
+      });
+
       this.updateFailure = false;
       this.confirmHeading = 'Cancel Recurring Donation?';
       this.confirmBody = `<p>By selecting <b>yes</b>, you are canceling your recurring donation
@@ -320,6 +179,11 @@ export default {
           await this[USER_TYPES.closeRdo]({
             rdoId: rdo.id,
             stripeSubscriptionId: rdo.stripe_subscription_id,
+          });
+          window.dataLayer.push({
+            event: 'customUserPortalCancelSuccess',
+            gaAction: this.ga.userPortal.actions['successful-cancel-donation'],
+            gaLabel: this.ga.userPortal.labels['cancel-donation'],
           });
           this.messageHeading='Donation Cancelled Successfully';
           this.messageBody = `<div class="t-size-s">Recurring donation of $${rdo.amount} (${rdo.period}) has been cancelled.</div>
@@ -342,15 +206,14 @@ export default {
             if (
               err.extra.data.detail === "missing data"
             ) {
-              this.failureMessage = 'An internal error occurred. Please try again and if the issue persists, contact us at membership@texastribune.org'
+              this.messageBody = 'An internal error occurred. Please try again and if the issue persists, contact us at membership@texastribune.org'
             }
             else if (
               err.extra.data.detail === "invalid card"
             ) {
-              this.failureMessage = 'The submitted card was declined or invalid. Please check your information and resubmit'
+              this.messageBody = 'The submitted card was declined or invalid. Please check your information and resubmit'
             }
             this.messageHeading = "Donation Cancelation Failed";
-            this.messageBody = this.failureMessage;
             this.messageType = 'failure';
           }
         }
